created KidProfileTag (like initials from iPhone contacts; change to pic if desired); added transactions refresh button (hot fix for refreshing all transactions on profile after creation); cleaned up some code in ProfileView and added Utilities;
