+++ piggy_bankers_app.xcodeproj/xcuserdata/mitchrogers.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -70,7 +70,7 @@
 		<key>piggy_bankers_app.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>1</integer>
+			<integer>0</integer>
 		</dict>
 	</dict>
 </dict>
