+++ piggy_bankers_app/View/Profile/AllTransactionsRow.swift
@@ -11,18 +11,26 @@ struct AllTransactionsRow: View {
     let transaction: Transaction
     
     var body: some View {
-        VStack(alignment: .leading) {
-            Text("\(NumberFormatter.localizedString(from: transaction.transaction_amount as NSNumber, number: .currency))")
-                .foregroundColor(transaction.transaction_type == "Sent" ? .green : .red)
-            Text("\(transaction.transaction_memo)")
-                .foregroundColor(.secondary)
-        }
-        
-        Spacer()
-        
-        VStack(alignment: .trailing) {
-            Text("\(transaction.transaction_date.formatted(.dateTime.day().month().year()))")
-                .foregroundColor(.secondary)
+        NavigationLink {
+            TransactionDetails(transaction: transaction)
+        } label: {
+            HStack() {
+                KidProfileTag(profile_id: String(transaction.profile_id))
+                
+                VStack(alignment: .leading) {
+                    Text("\(NumberFormatter.localizedString(from: transaction.transaction_amount as NSNumber, number: .currency))")
+                        .foregroundColor(transaction.transaction_type == "Sent" ? .green : .red)
+                    Text("\(transaction.transaction_memo)")
+                        .foregroundColor(.secondary)
+                }
+                
+                Spacer()
+                
+                VStack(alignment: .trailing) {
+                    Text("\(transaction.transaction_date.formatted(.dateTime.day().month().year()))")
+                        .foregroundColor(.secondary)
+                }
+            }
         }
     }
 }
