+++ piggy_bankers_app/ViewModel/Service/TransactionService.swift
@@ -8,7 +8,7 @@
 import Foundation
 import PostgresClientKit
 
-struct TransactionService {
+class TransactionService: ObservableObject {
     
     static func createTransaction(transaction: Transaction) {
         let newTransaction = transaction
@@ -23,7 +23,7 @@ struct TransactionService {
             let connection = try PostgresClientKit.Connection(configuration: configuration)
 //            defer { connection.close() }
             
-            let query = "INSERT INTO app.transaction VALUES (103, \(newTransaction.household_id), \(newTransaction.profile_id), '\(newTransaction.transaction_date)', '\(newTransaction.transaction_type)', \(newTransaction.transaction_amount), '\(newTransaction.transaction_memo)', '\(newTransaction.transaction_description)');"
+            let query = "INSERT INTO app.transaction VALUES (\(newTransaction.profile_id), \(newTransaction.household_id), \(newTransaction.profile_id), '\(newTransaction.transaction_date)', '\(newTransaction.transaction_type)', \(newTransaction.transaction_amount), '\(newTransaction.transaction_memo)', '\(newTransaction.transaction_description)');"
             
             let statement = try connection.prepareStatement(text: query)
             defer { statement.close() }
@@ -34,8 +34,6 @@ struct TransactionService {
         } catch {
             print("error creating record in db.transaction: \(error)")
         }
-        
-        fetchTransactionsPG()
     }
     
     static func fetchTransactionsPG() -> [Transaction]? {
@@ -80,25 +78,10 @@ struct TransactionService {
                 let transaction_amount = try columns[transaction_amount_col_num].double()
                 let transaction_memo = try columns[transaction_memo_col_num].string()
                 let transaction_description = try columns[transaction_description_col_num].string()
-                print("\(transaction_amount) - \(transaction_date)")
                 
-                let isoDate = transaction_date
-                print(isoDate)
-
-                let dateFormatter = DateFormatter()
-                dateFormatter.locale = Locale(identifier: "en_US") // set locale to reliable US_POSIX
-                dateFormatter.dateFormat = "yyyy-MM-dd HH:mm:ss"
-                var dateOf = Date()
-                if let date = dateFormatter.date(from: isoDate) {
-                    dateOf = date
-                }
-                
-                print(dateOf)
-
-                let calendar = Calendar.current
-                let components = calendar.dateComponents([.year, .month, .day], from: dateOf)
-                
-                let finalDate: Date = calendar.date(from: components) ?? Date()
+                // transaction_date from PG is stored as String. shortDate() is an extension of String that converts String to Date.
+                // String regex must match "yyyy-MM-dd HH:mm:ss"
+                let finalDate: Date = transaction_date.shortDate()
                 
                 allTransactions.append(Transaction(household_id: household_id, profile_id: profile_id, transaction_date: finalDate, transaction_type: transaction_type, transaction_amount: transaction_amount, transaction_memo: transaction_memo, transaction_description: transaction_description))
             }
